--- a/paddlenlp/ops/faster_transformer/src/fusion_gpt_op.h
+++ b/paddlenlp/ops/faster_transformer/src/fusion_gpt_op.h
@@ -3,9 +3,9 @@
 #include <string>
 #include <vector>
 
-#include "fastertransformer/common.h"
 #include "fastertransformer/gpt2.h"
 #include "fastertransformer/open_decoder.h"
+#include "fastertransformer/utils/common.h"
 
 #ifdef PADDLE_ON_INFERENCE
 #include "paddle/include/experimental/ext_all.h"