--- a/paddlenlp/ops/faster_transformer/transformer/decoding.py
+++ b/paddlenlp/ops/faster_transformer/transformer/decoding.py
@@ -280,8 +280,7 @@ class InferTransformerDecoding(nn.Layer):
             decoder.norm.bias = transfer_param(decoder.norm.bias, is_bias=True)
 
             linear.weight = transfer_param(linear.weight)
-            if "beam_search" != decoding_strategy:
-                linear.bias = transfer_param(linear.bias)
+            linear.bias = transfer_param(linear.bias)
 
             positional_embedding.weight = transfer_param(
                 positional_embedding.weight)