--- a/paddlenlp/ops/patches/FasterTransformer/utils/allocator.h
+++ b/paddlenlp/ops/patches/FasterTransformer/utils/allocator.h
@@ -22,8 +22,8 @@
 
 #include <cuda_runtime.h>
 #include <vector>
-#include "fastertransformer/common.h"
-#include "fastertransformer/utils.h"
+#include "fastertransformer/utils/common.h"
+#include "fastertransformer/utils/utils.h"
 
 #ifdef PADDLE_CUDA
 #ifdef PADDLE_ON_INFERENCE