… paddlenlp/ops/patches/FasterTransformer/common.h
… paddlenlp/ops/patches/FasterTransformer/cuda/topk_kernels.cu