<!-- Demo: https://github.com/PaddlePaddle/PaddleNLP/pull/26 -->
### PR types
<!-- One of [ New features | Bug fixes | Function optimization | Performance optimization | Breaking changes | Others ] -->
Breaking changes

### PR changes
<!-- One of [ Models | APIs | Docs | Others ] -->
APIs

### Description
<!-- Describe what this PR does -->
Promote to v4.0. Should review after #569 

TODOS:
* promote GPT
* remove `nccl` and `mpi` support in GPT. 
* topk sampling stochasticity verification. 
* #569 support.
* more ...