@FrostML 2021-07-02
**Transformer v4.0 Performace:** 
| hyper params | FT v3.1 fp32(in ms/batch) | FT v3.1 fp16(in ms/batch) | Original(in ms/batch) | FT v3.1 speedup(compare with fp32) | FT 4.0 fp32(in ms/batch) | FT v4.0 fp16(in ms/batch) | FT v4.0 speedup(compare with fp32) |
|--|--|--|--|--|--|--|--|
| batch_size: 1, beam_size: 1 | 36.6575 | 41.9541 | 1288.0803@enlei | 35.1382@enlei |  |  |  |
| batch_size: 1, beam_size: 4 | 49.5412 | 46.2929 | 1232.0821 | 24.87 |  |  |  |
| batch_size: 8, beam_size: 1 | 67.2444 | 64.3771 | 1997.2295 | 29.70 |  |  |  |
| batch_size: 8, beam_size: 4 | 75.6515 | 66.5741 | 2135.4212 | 28.23 |  |  |  |
| batch_size: 32, beam_size: 1 | 97.9693 | 84.9347 | 2633.8843 | 26.88 |  |  |  |
| batch_size: 32, beam_size: 4 | 184.9824 | 133.9078 | 2606.9408 | 14.09 |  |  |  |
| batch_size: 64, beam_size: 1 | 152.0229 | 122.5850 | 2881.6411 | 18.96 | 139.2874 | 113.4201 |  |
| batch_size: 64, beam_size: 4 | 347.8825 | 218.8199 | 3391.2002 | 9.75 | 271.1225 | 169.0535 |  |