@ZeyuChen 2021-07-03
@FrostML 看上去性能结论，V4.0 相比 V3.1 没有明显提升，某些场景有下滑。同时带来了依赖增多编译复杂的问题？
可否对优劣势做下总结？